:hammer: Refactor file naming and split classes
