+++ app/src/main/java/network/omisego/omgwallet/databinding/ImageViewBinding.kt
@@ -6,7 +6,7 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.bumptech.glide.request.RequestOptions
-import network.omisego.omgwallet.extensions.dpToPx
+import network.omisego.omgwallet.extension.dpToPx
 
 /*
  * OmiseGO
@@ -15,7 +15,7 @@ import network.omisego.omgwallet.extensions.dpToPx
  * Copyright © 2017-2018 OmiseGO. All rights reserved.
  */
 
-object ImageViewUtil {
+object ImageViewBinding {
     @JvmStatic
     @BindingAdapter("loadImage")
     fun loadImage(view: ImageView, imageUrl: String) {
