+++ app/src/main/java/network/omisego/omgwallet/extension/AndroidVersionExt.kt
@@ -1,4 +1,4 @@
-package network.omisego.omgwallet.extensions
+package network.omisego.omgwallet.extension
 
 /*
  * OmiseGO
