+++ app/src/main/java/network/omisego/omgwallet/extension/LiveDataExt.kt
@@ -1,15 +1,27 @@
-package network.omisego.omgwallet.util
+package network.omisego.omgwallet.extension
 
 /*
  * OmiseGO
  *
- * Created by Phuchit Sirimongkolsathien on 13/8/2018 AD.
+ * Created by Phuchit Sirimongkolsathien on 12/8/2018 AD.
  * Copyright © 2017-2018 OmiseGO. All rights reserved.
  */
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Transformations
+
+fun <T> mutableLiveDataOf(initial: T? = null): MutableLiveData<T> {
+    return MutableLiveData<T>().apply { this.value = initial }
+}
+
+fun <T> MutableLiveData<T>.fetchedThenCache(fetch: (MutableLiveData<T>) -> LiveData<T>): LiveData<T> =
+    if (value != null) {
+        this
+    } else {
+        fetch(this)
+    }
 
 fun <T, R> LiveData<T>.map(transform: (T) -> R) = Transformations.map(this, transform)
