--- app/src/main/java/network/omisego/omgwallet/extensions/Px.kt
@@ -1,15 +1,0 @@
-package network.omisego.omgwallet.extensions
-
-import android.content.Context
-import android.util.TypedValue
-
-/*
- * OmiseGO
- *
- * Created by Phuchit Sirimongkolsathien on 20/8/2018 AD.
- * Copyright © 2017-2018 OmiseGO. All rights reserved.
- */
-
-fun Context.dpToPx(dp: Float): Int {
-    return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, resources.displayMetrics).toInt()
-}
