+++ app/src/main/java/network/omisego/omgwallet/listener/MinimalTextChangedListener.kt
@@ -1,4 +1,4 @@
-package network.omisego.omgwallet.util
+package network.omisego.omgwallet.listener
 
 /*
  * OmiseGO
