+++ app/src/main/java/network/omisego/omgwallet/pages/signin/BiometricCallback.kt
@@ -1,4 +1,4 @@
-package network.omisego.omgmerchant.pages.signin
+package network.omisego.omgwallet.pages.signin
 
 import android.hardware.biometrics.BiometricPrompt
 import android.os.Build.VERSION_CODES.P
