+++ app/src/main/java/network/omisego/omgwallet/validator/PasswordValidator.kt
@@ -1,0 +1,33 @@
+package network.omisego.omgwallet.validator
+
+/*
+ * OmiseGO
+ *
+ * Created by Phuchit Sirimongkolsathien on 18/9/2018 AD.
+ * Copyright © 2017-2018 OmiseGO. All rights reserved.
+ */
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.Observer
+import network.omisego.omgwallet.model.ValidateResult
+
+class PasswordValidator(override var byPass: LiveData<Boolean>) : Validator(byPass) {
+    private val isPasswordLessThanEight: (String) -> Boolean = {
+        it.length < 8
+    }
+
+    override fun check(text: String, updateUI: ((ValidateResult) -> Unit)?): ValidateResult {
+        this.updateUI = updateUI
+        this.recentText = text
+        validation = when {
+            byPass.value == false && isPasswordLessThanEight(text) -> ValidateResult(false, "Password must contain at least 8 characters")
+            else -> ValidateResult(true)
+        }
+        updateUI?.invoke(validation)
+        return validation
+    }
+
+    init {
+        byPass.observe(this, Observer { check(this.recentText, this.updateUI) })
+    }
+}
